@@ -68,26 +68,35 @@ module Sass::Script::Value
       Number.new(number, *parse_unit_string(unit_string))
     end
 
-    # @overload list(*elements, separator)
+    # @overload list(*elements, separator:, bracketed: false)
     #   Create a space-separated list from the arguments given.
     #   @param elements [Array<Sass::Script::Value::Base>] Each argument will be a list element.
     #   @param separator [Symbol] Either :space or :comma.
+    #   @param bracketed [Boolean] Whether the list uses square brackets.
     #   @return [Sass::Script::Value::List] The space separated list.
     #
-    # @overload list(array, separator)
+    # @overload list(array, separator:, bracketed: false)
     #   Create a space-separated list from the array given.
     #   @param array [Array<Sass::Script::Value::Base>] A ruby array of Sass values
     #     to make into a list.
+    #   @param separator [Symbol] Either :space or :comma.
+    #   @param bracketed [Boolean] Whether the list uses square brackets.
     #   @return [Sass::Script::Value::List] The space separated list.
-    def list(*elements)
-      unless elements.last.is_a?(Symbol)
-        raise ArgumentError.new("A list type of :space or :comma must be specified.")
+    def list(*elements, separator: nil, bracketed: false)
+      # Support passing separator as the last value in elements for
+      # backwards-compatibility.
+      if separator.nil?
+        if elements.last.is_a?(Symbol)
+          separator = elements.pop
+        else
+          raise ArgumentError.new("A separator of :space or :comma must be specified.")
+        end
       end
-      separator = elements.pop
+
       if elements.size == 1 && elements.first.is_a?(Array)
         elements = elements.first
       end
-      Sass::Script::Value::List.new(elements, separator)
+      Sass::Script::Value::List.new(elements, separator: separator, bracketed: bracketed)
     end
 
     # Construct a Sass map.
