@@ -12,7 +12,7 @@ module Sass::Script::Value
     #
     # @param hash [Hash<Node, Node>]
     def initialize(hash)
-      super(Sass::Util.ordered_hash(hash))
+      super(hash)
     end
 
     # @see Value#options=
@@ -32,7 +32,7 @@ module Sass::Script::Value
     # @see Value#to_a
     def to_a
       value.map do |k, v|
-        list = List.new([k, v], :space)
+        list = List.new([k, v], separator: :space)
         list.options = options
         list
       end
