@@ -32,7 +32,7 @@ module Sass
       # @return [String] The escaped character
       # @private
       def self.escape_char(c)
-        return "\\%06x" % (Sass::Util.ord(c)) unless c =~ %r{[ -/:-~]}
+        return "\\%06x" % c.ord unless c =~ %r{[ -/:-~]}
         "\\#{c}"
       end
 
@@ -50,13 +50,7 @@ module Sass
       H        = /[0-9a-fA-F]/
       NL       = /\n|\r\n|\r|\f/
       UNICODE  = /\\#{H}{1,6}[ \t\r\n\f]?/
-      s = if Sass::Util.ruby1_8?
-            '\200-\377'
-          elsif Sass::Util.macruby?
-            '\u0080-\uD7FF\uE000-\uFFFD\U00010000-\U0010FFFF'
-          else
-            '\u{80}-\u{D7FF}\u{E000}-\u{FFFD}\u{10000}-\u{10FFFF}'
-          end
+      s = '\u{80}-\u{D7FF}\u{E000}-\u{FFFD}\u{10000}-\u{10FFFF}'
       NONASCII = /[#{s}]/
       ESCAPE   = /#{UNICODE}|\\[ -~#{s}]/
       NMSTART  = /[_a-zA-Z]|#{NONASCII}|#{ESCAPE}/
@@ -127,13 +121,13 @@ module Sass
       OPTIONAL = /!#{W}optional/i
       IDENT_START = /-|#{NMSTART}/
 
-      IDENT_HYPHEN_INTERP = /-(#\{)/
+      IDENT_HYPHEN_INTERP = /-(?=#\{)/
       STRING1_NOINTERP = /\"((?:[^\n\r\f\\"#]|#(?!\{)|#{ESCAPE})*)\"/
       STRING2_NOINTERP = /\'((?:[^\n\r\f\\'#]|#(?!\{)|#{ESCAPE})*)\'/
       STRING_NOINTERP = /#{STRING1_NOINTERP}|#{STRING2_NOINTERP}/
 
       STATIC_COMPONENT = /#{IDENT}|#{STRING_NOINTERP}|#{HEXCOLOR}|[+-]?#{NUMBER}|\!important/i
-      STATIC_VALUE = %r(#{STATIC_COMPONENT}(\s*[\s,/]\s*#{STATIC_COMPONENT})*([;}]))i
+      STATIC_VALUE = %r(#{STATIC_COMPONENT}(\s*[\s,\/]\s*#{STATIC_COMPONENT})*(?=[;}]))i
       STATIC_SELECTOR = /(#{NMCHAR}|[ \t]|[,>+*]|[:#.]#{NMSTART}){1,50}([{])/i
     end
   end
