@@ -282,7 +282,7 @@ SCSS
 
     _, sourcemap = engine.render_with_sourcemap('http://1.example.com/style.map')
 
-    uri = Sass::Util.file_uri_from_path(Sass::Util.absolute_path(filename_for_test(:scss)))
+    uri = Sass::Util.file_uri_from_path(File.absolute_path(filename_for_test(:scss)))
     assert_equal <<JSON.strip, sourcemap.to_json(:css_uri => 'css_uri')
 {
 "version": 3,
@@ -309,7 +309,7 @@ SCSS
 
     _, sourcemap = engine.render_with_sourcemap('http://1.example.com/style.map')
 
-    uri = Sass::Util.file_uri_from_path(Sass::Util.absolute_path(filename_for_test(:scss)))
+    uri = Sass::Util.file_uri_from_path(File.absolute_path(filename_for_test(:scss)))
     assert_equal <<JSON.strip, sourcemap.to_json(:css_uri => 'css_uri', :css_path => 'css_path')
 {
 "version": 3,
